requirements.txt
# Deep Learning Framework
torch
torchvision

# Web Application
streamlit

# Data Processing & Visualization
pandas
numpy
matplotlib
scikit-learn
Pillow
tqdm
jupyter